+++ src/pages/product/product-page.js
@@ -1,13 +1,15 @@
 import React, { Component } from 'react';
 
-class product extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {  }
-    }
-    render() { 
-        return (  );
-    }
+class ProductPage extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {  }
+  }
+  render() { 
+    return (  
+      <h1>This is products page</h1>
+    );
+  }
 }
  
-export default product-page;
+export default ProductPage;
